digital-twin/requirements.txt
scikit-learn
joblib
pandas
google-generativeai
python-dotenv
googlesearch-python
requests
beautifulsoup4
aiohttp
jinja2
langgraph
langchain
langchain_google_genai
playwright
playwright-stealth
httpx
pytest
fastapi
